backend/requirements.txt
Django>=5.0
djangorestframework>=3.15
django-cors-headers>=4.3
django-filter>=24.0
python-decouple>=3.8
